flask_jwt_extended feature has been added

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource,reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required,get_jwt_claims,jwt_optional,get_jwt_identity,fresh_jwt_required
 from models.item import ItemModel
 
 class Item(Resource):
@@ -7,13 +7,14 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="this field cannot be left blank")
     parser.add_argument('store_id', type=int, required=True, help="Every item needs a store id")
 
-    @jwt_required()
+    @jwt_required
     def get(self,name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {"message":"Item not found"}
 
+    @fresh_jwt_required
     def post(self,name):
         if ItemModel.find_by_name(name):
             return {"message":"Item already exists"},400
@@ -23,7 +24,12 @@ class Item(Resource):
         item.save_to_db()
         return item.json(),201
 
+    @jwt_required
     def delete(self,name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message":"Admin privilege required"}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -41,6 +47,14 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"items":[item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items":items}
+        return {"items":[item['name'] for item in items],
+                "message":"more data available if you log in"
+                }
+
